requirements.txt
requests[socks]
python-telegram-bot
toml
